moved logging tests to one file for each of curriculum_tracking and git_real apps
